+++ project/include/Enemy.h
@@ -4,17 +4,17 @@
 #include "Entity.h"
 #include "Player.h"
 
-class Enemy: public Entity {
+class Enemy : public Entity {
  public:
   explicit Enemy(float x, float y, int height, int weight);
   void Update(float time, std::vector<Object> &obj) override;
-  int GetDmg() const;
-  void TakeDmg(int getDmg);
+  float GetDmg() const;
+  void TakeDmg(float getDmg);
   float GetTimer() const;
   void ResetTimer();
 
  private:
-  int dmg;
-  int hp;
+  float dmg;
+  float hp;
   float timerHit;
 };
